Support for strategic-merge
This change, if applied, adds strategic-merge custom merge strategy to
agnosticv.

It can be applied to dict or list using the x-merge key in any schema.

Once this change is merged, we will update in a separate PR the default for `__meta__`
and `agnosticv_meta` variables to use 'strategic-merge' instead of 'merge'.