Add option to load related file into merged vars
see GPTEINFRA-5776 and GPTEINFRA-5671

Add a new version of the 'related_files' entry in configuration file: 'related_file_v2'.

The option provides a way to load the content of the file into a path.

The path is JSON pointer format, like the other paths in the configuration (x-merge path
in schema file.)

Backward compatibility: Adding a new version of 'related_files' entry
will ensure the previous one still works as expected
across the different versions of the agV repos.
Another solution would have been to support both list of strings and
list of maps, but that would make the definition of the configuration less clear.
We can also retire v1 later when it's been removed from all repos and
not used anymore.