Perf improvement: process '-has' flag at the end
The '--has' flag is costly because it forces variables to be merged.
When looping over a lot of catalog items, it can slow down the command a
lot.

Move its processing at the last position when filtering catalog items, so it's
executed less often when used in combination of other flags.